Add perturbations from turbsim
to TVMInletOutlet refValue (i.e., the inletValue).

This ended up being a massive PITA because although the _apparent_ Type
is "vector", as expected for field U and output when printing the
complete fvPatchField dictionary to screen, the _actual_ Type ends up
either being (or looking like) a Foam::SphericalTensor. So we cannot
cleanly loop over all the faces (e.g., those of patch().Cf()), or access
and modify the components with .x(), .y(), and .z() accessors. In
addition, we cannot simply add on the perturbation field with a
refValue() += perts statement. This creates additional issues such as
when creating the perturbations field, which should be a Field<Type=vector>,
there's a type mismatch problem because the actual type is something
tensor-like.

The workaround is to directly retrieve a _copy_ of all x-, y-, and z-
components with refValue().component(d) (where d=0,1,2) and then update
it using refValue().replace(d, newRefValuesList).

Not sure if this is a bug or an odd implementation quirk.