Add -ref option to specify a sampled plane (foamFile format) to use as a reference
This is needed because the boundaryDataPre sampling that is output does not
have consistent ordering (e.g. C-indexing with z increasing the fastest and x
the slowest), so if we were to map the sampled inflow to a deformed mesh (e.g.,
generated with moveDynamicMesh to conform to terrain features), the data would
end up garbled if we had not reordered it.

Removed the -cellDisplacement option as this does not work as expected.

Lastly, changed the loop order to first write out all patches at each time so
that complete boundaryData time directories are available immediately.